Improve schema scan performance StarRocks#4028: change int32 to int64

+++ be/src/exec/vectorized/schema_scanner.h
@@ -31,7 +31,7 @@ struct SchemaScannerParam {
     const std::string* ip{nullptr};                   // frontend ip
     int32_t port{0};                                  // frontend thrift port
     int64_t thread_id = 0;
-    int32_t limit{0};
+    int64_t limit{0};
 
     RuntimeProfile::Counter* _rpc_timer = nullptr;
     RuntimeProfile::Counter* _fill_chunk_timer = nullptr;
